Fix texture buffer drawing in tealeaf_canvas
ctx->on_screen turns out to not be a good indicator of whether the
context is being drawn on screen or not. To prevent the context from
generating its own texture name, on_screen must be true. In the case of
external engines where a texture name is provided, ctx->on_screen must
be true. If the context is truly being drawn directly to screen, there
will be no destination texture.